Update 'first_run.sh' to 'quickstart.sh' (#704)
* Add virtualenv setup to initial commands

Note the instructions create a directory named 'env'.  I prefer 'venv'
because it's more distinctive, but this project's .gitignore is
already set up to ignore 'env', so let's use that.

* add /etc/hosts and browser URL to datagen output

* rename 'first_run.sh' to 'quickstart.sh'

+++ quickstart.sh
@@ -26,8 +26,14 @@ else
 fi
 
 
-# install basic requirements as the local user
-pip3 install --user -r requirements.txt
+# install basic requirements either in venv or as the local user
+if test -f env/bin/activate
+then
+	source env/bin/activate;
+	pip3 install -r requirements.txt;
+else
+	pip3 install --user -r requirements.txt;
+fi
 echo $SPACER
 ./fetch-vendor-resources.sh
